requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
crewai>=0.28.0,<0.100.0
chromadb>=0.4.0
flashrank>=0.2.0
tavily-python>=0.3.0
openai>=1.0.0

# 测试依赖
pytest>=7.0.0
pytest-cov>=4.0.0
